+++ moveit_ros/visualization/rviz_plugin_render_tools/include/moveit/rviz_plugin_render_tools/planning_scene_render.h
@@ -70,7 +70,7 @@ public:
   void renderPlanningScene(const planning_scene::PlanningSceneConstPtr& scene,
                            const Ogre::ColourValue& default_scene_color,
                            const Ogre::ColourValue& default_attached_color, OctreeVoxelRenderMode voxel_render_mode,
-                           OctreeVoxelColorMode voxel_color_mode, float default_scene_alpha);
+                           OctreeVoxelColorMode voxel_color_mode, double default_scene_alpha);
   void clear();
 
 private:
